+++ examples/maven-spring-preprocessor/java-example/pom.xml
@@ -32,7 +32,7 @@
                         <configuration>
                             <input>${project.basedir}/src/main/openapi/petstorev3.json</input>
                             <output>${project.build.directory}/generated-sources/openapi</output>
-                            <preProcessor>${project.basedir}/src/main/java/community/flock/wirespec/example/maven/preprocessor/ExamplePreProcessor.java</preProcessor>
+                            <preProcessor>${project.basedir}/src/main/java/community/flock/wirespec/example/maven/preprocessor/PreProcessor.java</preProcessor>
                             <format>OpenAPIV3</format>
                             <languages>
                                 <language>Java</language>
