--- examples/maven-spring-preprocessor/kotlin-example/src/main/kotlin/community/flock/wirespec/example/maven/preprocessor/ExamplePreProcessor.kt
@@ -1,6 +1,0 @@
-package community.flock.wirespec.example.maven.preprocessor
-
-class ExamplePreProcessor : (String) -> String {
-    override fun invoke(input: String): String = input
-}
-
