+++ src/plugin/maven/build.gradle.kts
@@ -1,7 +1,3 @@
-
-import org.w3c.dom.Node
-import org.w3c.dom.NodeList
-
 plugins {
     alias(libs.plugins.kotlin.jvm)
     alias(libs.plugins.maven.plugin.development)
@@ -37,21 +33,6 @@ publishing {
         register("mavenJava", MavenPublication::class) {
             artifactId = "wirespec-maven-plugin"
             from(components["java"])
-//            pom.withXml {
-//                val root = asElement()
-//                println("childNodes: ${root.childNodes.length}")
-//                val dependencies = root.childNodes.find {it.nodeName == "dependencies"}
-//                val kotlinCompilerDep = dependencies?.childNodes?.map {
-//                    val dep = it.childNodes.find { it.nodeName == "artifactId" && it.textContent == "kotlin-compiler-embeddable"}
-//                    if(dep != null){
-//                        val version = it.childNodes.find { it.nodeName == "version" }
-//                        version?.textContent = "${'$'}{kotlin.version}"
-//                    }
-//                    println(dep?.textContent)
-//                }
-//
-//
-//            }
         }
     }
 }
@@ -67,15 +48,3 @@ mavenPlugin {
     description.set("Plugin to run wirespec compiler")
     goalPrefix.set("wirespec")
 }
-
-private fun <T> NodeList.map(action: (Node) -> T) = (0 until length)
-    .map{item(it)}
-    .map {  action(it) }
-
-private fun NodeList.filter(predicate: (Node) -> Boolean) = (0 until length)
-    .map{item(it)}
-    .filter {  predicate(it) }
-
-private fun NodeList.find(predicate: (Node) -> Boolean) = (0 until length)
-    .map { item(it) }
-    .find {  predicate(it) }
